+++ examples/showoci/showoci.py
@@ -8,7 +8,7 @@
 # Supports Python 2.7 and above, Python 3 recommended
 # With Python 2, execute using - python showoci.py
 #
-# coding: utf-8
+# coding: utf-8 ```
 ##########################################################################
 # OCI Report Tool SHOWOCI:
 #